third_party/WebKit/LayoutTests/http/tests/loading/htxg/htxg-location-origin-trial.html

<!DOCTYPE html>
<title>Location of SignedHTTPExchange with Origin-Trial header</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="./resources/htxg-util.js"></script>
<body>
<script>
promise_test(async (t) => {
  await waitUntilDidFinishLoadForFrame;

  // The timestamp of the test HTXG file is "Apr 1 2018 00:00 UTC" and valid
  // until "Apr 8 2018 00:00 UTC".
  await setSignedExchangeVerificationTime(new Date("Apr 1 2018 00:01 UTC"));

  const event = await new Promise(async (resolve, reject) => {
    // We can't catch the network error on iframe. So we use the timer.
    t.step_timeout(() => reject('timeout'), 1000);

    const frame =
        await withIframe('resources/htxg-location-origin-trial.php',
                         'htxg_iframe');
    const channel = new MessageChannel();
    channel.port1.onmessage = resolve;
    frame.contentWindow.postMessage(
        {port: channel.port2}, '*', [channel.port2]);
  });
  assert_equals(event.data.location, 'https://www.127.0.0.1/test.html');
}, 'Location of SignedHTTPExchange with Origin-Trial header');
</script>
</body>
